make numba workable
