+++ graspologic/models/__init__.py
@@ -5,6 +5,13 @@ from .base import BaseGraphEstimator
 from .er import DCEREstimator, EREstimator
 from .rdpg import RDPGEstimator
 from .sbm_estimators import DCSBMEstimator, SBMEstimator
+from .edge_swaps import (
+    _do_setup,
+    _do_some_edge_swaps,
+    _numba_edge_swap,
+    _edge_swap,
+    _do_swap,
+)
 
 __all__ = [
     "BaseGraphEstimator",
@@ -13,4 +20,9 @@ __all__ = [
     "SBMEstimator",
     "DCSBMEstimator",
     "RDPGEstimator",
+    "_do_setup",
+    "_do_some_edge_swaps",
+    "_numba_edge_swap",
+    "_edge_swap",
+    "_do_swap",
 ]
