+++ graspologic/utils/utils.py
@@ -1156,206 +1156,6 @@ def remap_node_ids(
     return graph_remapped, node_id_dict
 
 
-def edge_swap_setup(adjacency):
-    if type(adjacency) == np.array or type(adjacency) == np.ndarray:
-        edge_list = []
-        row_inds, col_inds = np.nonzero(adjacency)
-        for m, n in zip(row_inds, col_inds):
-            edge_list.append([m, n])
-
-        edge_list = np.array(edge_list)
-        return adjacency, edge_list
-
-    elif type(adjacency) == lil_matrix:
-        edge_list = np.array(adjacency.nonzero())
-        return adjacency, edge_list
-
-
-def checks_swap(adjacency, edge_list: np.array):
-    if type(adjacency) == np.array or type(adjacency) == np.ndarray:
-        adjacency, edge_list = simple_checks_swap(adjacency, edge_list)
-        return adjacency, edge_list
-
-    elif type(adjacency) == lil_matrix:
-        # checks if there are at 2 edges in the graph
-        if adjacency.nnz < 2:
-            # print("graph has less than two edges")
-            return adjacency, edge_list
-
-        # choose two indices at random
-        rng = np.random.default_rng(np.random.randint(1234))
-        orig_inds = rng.choice(adjacency.nnz, size=2, replace=False)
-
-        u, v = edge_list[0][orig_inds[0]], edge_list[1][orig_inds[0]]
-        x, y = edge_list[0][orig_inds[1]], edge_list[1][orig_inds[1]]
-        # ensures no initial loops
-        if u == v or x == y:
-            # print("initial loops")
-            return adjacency, edge_list
-
-        # ensures no loops after swap
-        if u == x or v == y:
-            # print("loops after swap")
-            return adjacency, edge_list
-
-        # save edge values
-        w_uv = adjacency[u, v]
-        w_xy = adjacency[x, y]
-        w_ux = adjacency[u, x]
-        w_vy = adjacency[v, y]
-
-        # ensures no initial multigraphs
-        if w_uv > 1 or w_xy > 1:
-            # print("initial multigraph")
-            return adjacency, edge_list
-
-        # ensures no multigraphs after swap
-        if w_ux >= 1 or w_vy >= 1:
-            # print("multigraph after swap")
-            return adjacency, edge_list
-
-        # perform the swap
-        adjacency[u, v] = 0
-        adjacency[x, y] = 0
-
-        adjacency[u, x] = 1
-        adjacency[v, y] = 1
-
-        edge_list[0][orig_inds[0]], edge_list[1][orig_inds[0]] = u, x
-        edge_list[0][orig_inds[1]], edge_list[1][orig_inds[1]] = v, y
-        return adjacency, edge_list
-
-
-def simple_edge_swap_setup(adjacency: np.array):
-    # performs degree preserving edge swap on simple graph
-
-    # make edge list from adjacency matrix
-    edge_list = []
-    row_inds, col_inds = np.nonzero(adjacency)
-    for m, n in zip(row_inds, col_inds):
-        edge_list.append([m, n])
-
-    edge_list = np.array(edge_list)
-    return adjacency, edge_list
-
-
-def scipy_edge_swap_setup(adjacency: np.array):
-
-    # convert to lil_matrix
-    adjacency = lil_matrix(adjacency)
-    edge_list = np.array(adjacency.nonzero())
-
-    return adjacency, edge_list
-
-
-@jit(nopython=True, nogil=True)
-def simple_checks_swap(adjacency: np.array, edge_list: np.array):
-
-    # checks if there are at 2 edges in the graph
-    if len(edge_list) < 2:
-        # print("graph has less than two edges")
-        return adjacency, edge_list
-
-    # choose two indices at random
-    seed(np.random.randint(12345))
-    orig_inds = np.random.choice(len(edge_list), size=2, replace=False)
-
-    u, v = edge_list[orig_inds[0]]
-    x, y = edge_list[orig_inds[1]]
-    # ensures no initial loops
-    if u == v or x == y:
-        # print("initial loops")
-        return adjacency, edge_list
-
-    # ensures no loops after swap (must be swap on 4 distinct nodes)
-    if u == x or v == y:
-        # print("loops after swap")
-        return adjacency, edge_list
-
-    # save edge values
-    w_uv = adjacency[u, v]
-    w_xy = adjacency[x, y]
-    w_ux = adjacency[u, x]
-    w_vy = adjacency[v, y]
-
-    # ensures no initial multigraphs
-    if w_uv > 1 or w_xy > 1:
-        # print("initial multigraph")
-        return adjacency, edge_list
-
-    # ensures no multigraphs after swap
-    if w_ux >= 1 or w_vy >= 1:
-        # print("multigraph after swap")
-        return adjacency, edge_list
-
-    # perform the swap
-    adjacency[u, v] = 0
-    adjacency[x, y] = 0
-
-    adjacency[u, x] = 1
-    adjacency[v, y] = 1
-
-    # DO EDGE LIST STUFF
-    edge_list[orig_inds[0]] = [u, x]
-    edge_list[orig_inds[1]] = [v, y]
-    return adjacency, edge_list
-
-
-def scipy_checks_swap(adjacency: lil_matrix, edge_list: np.array):
-    # checks if there are at 2 edges in the graph
-    if adjacency.nnz < 2:
-        # print("graph has less than two edges")
-        return adjacency, edge_list
-
-    # choose two indices at random
-    rng = np.random.default_rng(np.random.randint(1234))
-    orig_inds = rng.choice(adjacency.nnz, size=2, replace=False)
-
-    u, v = edge_list[0][orig_inds[0]], edge_list[1][orig_inds[0]]
-    x, y = edge_list[0][orig_inds[1]], edge_list[1][orig_inds[1]]
-    # ensures no initial loops
-    if u == v or x == y:
-        # print("initial loops")
-        return adjacency, edge_list
-
-    # ensures no loops after swap
-    if u == x or v == y:
-        # print("loops after swap")
-        return adjacency, edge_list
-
-    # save edge values
-    w_uv = adjacency[u, v]
-    w_xy = adjacency[x, y]
-    w_ux = adjacency[u, x]
-    w_vy = adjacency[v, y]
-
-    # ensures no initial multigraphs
-    if w_uv > 1 or w_xy > 1:
-        # print("initial multigraph")
-        return adjacency, edge_list
-
-    # ensures no multigraphs after swap
-    if w_ux >= 1 or w_vy >= 1:
-        # print("multigraph after swap")
-        return adjacency, edge_list
-
-    # perform the swap
-    adjacency[u, v] = 0
-    adjacency[x, y] = 0
-
-    adjacency[u, x] = 1
-    adjacency[v, y] = 1
-
-    edge_list[0][orig_inds[0]], edge_list[1][orig_inds[0]] = u, x
-    edge_list[0][orig_inds[1]], edge_list[1][orig_inds[1]] = v, y
-    return adjacency, edge_list
-
-
-@njit
-def seed(a):
-    np.random.seed(a)
-
-
 def suppress_common_warnings() -> None:
     """
     Suppresses common warnings that occur when using graspologic.
